Yet another major refactoring
Yet again the compilation has been refactored. The
previous model did not pass input and output streams
around in a consistent fashion, leading to breakage
for closures.

This refactoring also makes commands hold state after
compilation, which leads to a major refactoring of
all commands. This leads to a slight increase in
type safety.